Move media player impl creator to own class
