+++ core/src/main/java/com/novoda/noplayer/mediaplayer/AndroidMediaPlayerImpl.java
@@ -1,10 +1,7 @@
 package com.novoda.noplayer.mediaplayer;
 
-import android.content.Context;
-import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Handler;
-import android.os.Looper;
 import android.view.SurfaceHolder;
 
 import com.novoda.noplayer.ContentType;
@@ -14,14 +11,11 @@ import com.novoda.noplayer.LoadTimeout;
 import com.novoda.noplayer.Player;
 import com.novoda.noplayer.PlayerAudioTrack;
 import com.novoda.noplayer.PlayerListenersHolder;
-import com.novoda.noplayer.PlayerState;
 import com.novoda.noplayer.PlayerView;
 import com.novoda.noplayer.SurfaceHolderRequester;
-import com.novoda.noplayer.SystemClock;
 import com.novoda.noplayer.Timeout;
 import com.novoda.noplayer.VideoDuration;
 import com.novoda.noplayer.VideoPosition;
-import com.novoda.noplayer.mediaplayer.forwarder.MediaPlayerForwarder;
 import com.novoda.noplayer.player.PlayerInformation;
 
 import java.util.List;
@@ -47,70 +41,6 @@ public final class AndroidMediaPlayerImpl implements Player {
     private boolean seekingWithIntentToPlay;
     private SurfaceHolderRequester surfaceHolderRequester;
 
-    public static AndroidMediaPlayerImpl newInstance(Context context) {
-        LoadTimeout loadTimeout = new LoadTimeout(new SystemClock(), new Handler(Looper.getMainLooper()));
-        MediaPlayerForwarder forwarder = new MediaPlayerForwarder();
-        AndroidMediaPlayerFacade facade = AndroidMediaPlayerFacade.newInstance(context, forwarder);
-        PlayerListenersHolder listenersHolder = new PlayerListenersHolder();
-        CheckBufferHeartbeatCallback bufferHeartbeatCallback = new CheckBufferHeartbeatCallback();
-        Heart heart = Heart.newInstance();
-        BuggyVideoDriverPreventer buggyVideoDriverPreventer = BuggyVideoDriverPreventer.newInstance();
-        Handler handler = new Handler(Looper.getMainLooper());
-
-        return newInstance(loadTimeout, forwarder, facade, listenersHolder, bufferHeartbeatCallback, heart, buggyVideoDriverPreventer, handler);
-    }
-
-    static AndroidMediaPlayerImpl newInstance(LoadTimeout loadTimeout,
-                                              MediaPlayerForwarder forwarder,
-                                              AndroidMediaPlayerFacade facade,
-                                              PlayerListenersHolder listenersHolder,
-                                              CheckBufferHeartbeatCallback bufferHeartbeatCallback,
-                                              Heart heart,
-                                              BuggyVideoDriverPreventer preventer,
-                                              Handler handler) {
-        AndroidMediaPlayerImpl mediaPlayer = new AndroidMediaPlayerImpl(facade, listenersHolder, loadTimeout, heart, handler, preventer);
-        bindListeners(loadTimeout, forwarder, facade, listenersHolder, bufferHeartbeatCallback, heart, mediaPlayer);
-
-        return mediaPlayer;
-    }
-
-    private static void bindListeners(final LoadTimeout loadTimeout,
-                                      MediaPlayerForwarder forwarder,
-                                      final AndroidMediaPlayerFacade facade,
-                                      PlayerListenersHolder listenersHolder,
-                                      CheckBufferHeartbeatCallback bufferHeartbeatCallback,
-                                      Heart heart,
-                                      final AndroidMediaPlayerImpl mediaPlayer) {
-        forwarder.bind(listenersHolder.getPreparedListeners(), mediaPlayer);
-        forwarder.bind(listenersHolder.getBufferStateListeners(), listenersHolder.getErrorListeners(), mediaPlayer);
-        forwarder.bind(listenersHolder.getCompletionListeners(), listenersHolder.getStateChangedListeners());
-        forwarder.bind(listenersHolder.getVideoSizeChangedListeners());
-        forwarder.bind(listenersHolder.getInfoListeners());
-        bufferHeartbeatCallback.bind(forwarder.onHeartbeatListener());
-        heart.bind(new Heart.Heartbeat<>(listenersHolder.getHeartbeatCallbacks(), mediaPlayer));
-        listenersHolder.addHeartbeatCallback(bufferHeartbeatCallback);
-        listenersHolder.addPreparedListener(new PreparedListener() {
-            @Override
-            public void onPrepared(PlayerState playerState) {
-                loadTimeout.cancel();
-                facade.setOnSeekCompleteListener(mediaPlayer.seekToResettingSeekListener);
-            }
-        });
-        listenersHolder.addErrorListener(new ErrorListener() {
-            @Override
-            public void onError(Player player, PlayerError error) {
-                loadTimeout.cancel();
-            }
-        });
-        listenersHolder.addVideoSizeChangedListener(new VideoSizeChangedListener() {
-            @Override
-            public void onVideoSizeChanged(int width, int height, int unappliedRotationDegrees, float pixelWidthHeightRatio) {
-                mediaPlayer.videoWidth = width;
-                mediaPlayer.videoHeight = height;
-            }
-        });
-    }
-
     AndroidMediaPlayerImpl(AndroidMediaPlayerFacade mediaPlayer,
                            PlayerListenersHolder listenersHolder,
                            LoadTimeout loadTimeoutParam,
@@ -124,18 +54,6 @@ public final class AndroidMediaPlayerImpl implements Player {
         this.handler = handler;
         this.buggyVideoDriverPreventer = buggyVideoDriverPreventer;
     }
-
-    private final MediaPlayer.OnSeekCompleteListener seekToResettingSeekListener = new MediaPlayer.OnSeekCompleteListener() {
-        @Override
-        public void onSeekComplete(MediaPlayer mp) {
-            seekToPosition = NO_SEEK_TO_POSITION;
-
-            if (seekingWithIntentToPlay || isPlaying()) {
-                seekingWithIntentToPlay = false;
-                play();
-            }
-        }
-    };
 
     @Override
     public void play() {
@@ -190,9 +108,23 @@ public final class AndroidMediaPlayerImpl implements Player {
         surfaceHolderRequester.requestSurfaceHolder(callback);
     }
 
+    void resetSeek() {
+        seekToPosition = NO_SEEK_TO_POSITION;
+
+        if (seekingWithIntentToPlay || isPlaying()) {
+            seekingWithIntentToPlay = false;
+            play();
+        }
+    }
+
     private void seekWithIntentToPlay(VideoPosition position) {
         seekingWithIntentToPlay = true;
         seekTo(position);
+    }
+
+    void updateVideoSize(int width, int height) {
+        this.videoWidth = width;
+        this.videoHeight = height;
     }
 
     @Override
