+++ core/src/test/java/com/novoda/noplayer/mediaplayer/AndroidMediaPlayerImplTest.java
@@ -1,6 +1,5 @@
 package com.novoda.noplayer.mediaplayer;
 
-import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Handler;
@@ -25,7 +24,6 @@ import com.novoda.noplayer.listeners.InfoListeners;
 import com.novoda.noplayer.listeners.PreparedListeners;
 import com.novoda.noplayer.listeners.StateChangedListeners;
 import com.novoda.noplayer.listeners.VideoSizeChangedListeners;
-import com.novoda.noplayer.mediaplayer.forwarder.MediaPlayerForwarder;
 import com.novoda.noplayer.player.PlayerInformation;
 import com.novoda.noplayer.player.PlayerType;
 import com.novoda.notils.logger.simple.Log;
@@ -50,12 +48,7 @@ import org.mockito.stubbing.Answer;
 import static org.fest.assertions.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.doAnswer;
-import static org.mockito.Mockito.eq;
-import static org.mockito.Mockito.inOrder;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.*;
 
 @RunWith(Enclosed.class)
 public class AndroidMediaPlayerImplTest {
@@ -67,10 +60,6 @@ public class AndroidMediaPlayerImplTest {
 
         private static final long TWO_MINUTES_IN_MILLIS = 120000;
 
-        private static final int WIDTH = 100;
-        private static final int HEIGHT = 200;
-        private static final int ANY_ROTATION_DEGREES = 0;
-        private static final int ANY_PIXEL_WIDTH_HEIGHT = 1;
         private static final int ONE_SECOND_IN_MILLIS = 1000;
         private static final boolean IS_PLAYING = true;
         private static final PlayerAudioTrack PLAYER_AUDIO_TRACK = new PlayerAudioTrack(0, 0, "id", "english", ".mp4", 1, 120);
@@ -84,15 +73,11 @@ public class AndroidMediaPlayerImplTest {
         @Mock
         private PlayerListenersHolder listenersHolder;
         @Mock
-        private MediaPlayerForwarder forwarder;
-        @Mock
         private LoadTimeout loadTimeout;
         @Mock
         private Heart heart;
         @Mock
         private Handler handler;
-        @Mock
-        private CheckBufferHeartbeatCallback bufferHeartbeatCallback;
         @Mock
         private BuggyVideoDriverPreventer preventer;
         @Mock
@@ -109,17 +94,7 @@ public class AndroidMediaPlayerImplTest {
         private InfoListeners infoListeners;
         @Mock
         private StateChangedListeners stateChangedListeners;
-        @Mock
-        private MediaPlayer.OnPreparedListener onPreparedListener;
-        @Mock
-        private MediaPlayer.OnCompletionListener onCompletionListener;
-        @Mock
-        private MediaPlayer.OnErrorListener onErrorListener;
-        @Mock
-        private MediaPlayer.OnVideoSizeChangedListener onSizeChangedListener;
 
-        @Mock
-        private CheckBufferHeartbeatCallback.BufferListener bufferListener;
         private AndroidMediaPlayerImpl player;
 
         @Before
@@ -133,82 +108,7 @@ public class AndroidMediaPlayerImplTest {
             given(listenersHolder.getInfoListeners()).willReturn(infoListeners);
             given(listenersHolder.getStateChangedListeners()).willReturn(stateChangedListeners);
 
-            given(forwarder.onPreparedListener()).willReturn(onPreparedListener);
-            given(forwarder.onCompletionListener()).willReturn(onCompletionListener);
-            given(forwarder.onErrorListener()).willReturn(onErrorListener);
-            given(forwarder.onSizeChangedListener()).willReturn(onSizeChangedListener);
-            given(forwarder.onHeartbeatListener()).willReturn(bufferListener);
-
-            player = AndroidMediaPlayerImpl.newInstance(loadTimeout, forwarder, mediaPlayer, listenersHolder, bufferHeartbeatCallback, heart, preventer, handler);
-        }
-
-        @Test
-        public void whenCreatingAndroidMediaPlayerImpl_thenBindsListenersToForwarder() {
-            verify(forwarder).bind(preparedListeners, player);
-            verify(forwarder).bind(bufferStateListeners, errorListeners, player);
-            verify(forwarder).bind(completionListeners, stateChangedListeners);
-            verify(forwarder).bind(videoSizeChangedListeners);
-            verify(forwarder).bind(infoListeners);
-        }
-
-        @Test
-        public void whenCreatingAndroidMediaPlayerImpl_thenBindsListenerToBufferHeartbeatCallback() {
-            verify(bufferHeartbeatCallback).bind(bufferListener);
-        }
-
-        @Test
-        public void whenCreatingAndroidMediaPlayerImpl_thenBindsHeartbeatCallbackToListenerHolder() {
-            verify(listenersHolder).addHeartbeatCallback(bufferHeartbeatCallback);
-        }
-
-        @Test
-        public void givenBoundPreparedListener_whenCallingOnPrepared_thenCancelsTimeout() {
-            ArgumentCaptor<Player.PreparedListener> preparedListenerCaptor = ArgumentCaptor.forClass(Player.PreparedListener.class);
-            verify(listenersHolder).addPreparedListener(preparedListenerCaptor.capture());
-
-            Player.PreparedListener preparedListener = preparedListenerCaptor.getValue();
-            preparedListener.onPrepared(player);
-
-            verify(loadTimeout).cancel();
-        }
-
-        @Test
-        public void whenCreatingAndroidMediaPlayerImpl_thenBindsHeart() {
-            verify(heart).bind(any(Heart.Heartbeat.class));
-        }
-
-        @Test
-        public void givenBoundPreparedListener_whenCallingOnPrepared_thenSetsOnSeekCompleteListener() {
-            ArgumentCaptor<Player.PreparedListener> preparedListenerCaptor = ArgumentCaptor.forClass(Player.PreparedListener.class);
-            verify(listenersHolder).addPreparedListener(preparedListenerCaptor.capture());
-
-            Player.PreparedListener preparedListener = preparedListenerCaptor.getValue();
-            preparedListener.onPrepared(player);
-
-            verify(mediaPlayer).setOnSeekCompleteListener(any(MediaPlayer.OnSeekCompleteListener.class));
-        }
-
-        @Test
-        public void givenBoundErrorListener_whenCallingOnError_thenCancelsTimeout() {
-            ArgumentCaptor<Player.ErrorListener> errorListenerCaptor = ArgumentCaptor.forClass(Player.ErrorListener.class);
-            verify(listenersHolder).addErrorListener(errorListenerCaptor.capture());
-
-            Player.ErrorListener errorListener = errorListenerCaptor.getValue();
-            errorListener.onError(player, mock(Player.PlayerError.class));
-
-            verify(loadTimeout).cancel();
-        }
-
-        @Test
-        public void givenBoundVideoSizeChangedListener_whenCallingOnVideoSizeChanged_thenVideoWidthAndHeightMatches() {
-            ArgumentCaptor<Player.VideoSizeChangedListener> videoSizeChangedListenerCaptor = ArgumentCaptor.forClass(Player.VideoSizeChangedListener.class);
-            verify(listenersHolder).addVideoSizeChangedListener(videoSizeChangedListenerCaptor.capture());
-
-            Player.VideoSizeChangedListener videoSizeChangedListener = videoSizeChangedListenerCaptor.getValue();
-            videoSizeChangedListener.onVideoSizeChanged(WIDTH, HEIGHT, ANY_ROTATION_DEGREES, ANY_PIXEL_WIDTH_HEIGHT);
-
-            assertThat(player.getVideoWidth()).isEqualTo(WIDTH);
-            assertThat(player.getVideoHeight()).isEqualTo(HEIGHT);
+            player = new AndroidMediaPlayerImpl(mediaPlayer, listenersHolder, loadTimeout, heart, handler, preventer);
         }
 
         @Test
@@ -635,15 +535,11 @@ public class AndroidMediaPlayerImplTest {
         @Mock
         PlayerListenersHolder listenersHolder;
         @Mock
-        MediaPlayerForwarder forwarder;
-        @Mock
         LoadTimeout loadTimeout;
         @Mock
         Heart heart;
         @Mock
         Handler handler;
-        @Mock
-        CheckBufferHeartbeatCallback bufferHeartbeatCallback;
         @Mock
         BuggyVideoDriverPreventer buggyVideoDriverPreventer;
         @Mock
@@ -661,17 +557,7 @@ public class AndroidMediaPlayerImplTest {
         @Mock
         StateChangedListeners stateChangedListeners;
         @Mock
-        MediaPlayer.OnPreparedListener onPreparedListener;
-        @Mock
-        MediaPlayer.OnCompletionListener onCompletionListener;
-        @Mock
-        MediaPlayer.OnErrorListener onErrorListener;
-        @Mock
-        MediaPlayer.OnVideoSizeChangedListener onSizeChangedListener;
-        @Mock
         SurfaceHolder surfaceHolder;
-        @Mock
-        CheckBufferHeartbeatCallback.BufferListener bufferListener;
         @Mock
         PlayerView playerView;
         @Mock
@@ -704,12 +590,6 @@ public class AndroidMediaPlayerImplTest {
             given(listenersHolder.getVideoSizeChangedListeners()).willReturn(videoSizeChangedListeners);
             given(listenersHolder.getInfoListeners()).willReturn(infoListeners);
             given(listenersHolder.getStateChangedListeners()).willReturn(stateChangedListeners);
-
-            given(forwarder.onPreparedListener()).willReturn(onPreparedListener);
-            given(forwarder.onCompletionListener()).willReturn(onCompletionListener);
-            given(forwarder.onErrorListener()).willReturn(onErrorListener);
-            given(forwarder.onSizeChangedListener()).willReturn(onSizeChangedListener);
-            given(forwarder.onHeartbeatListener()).willReturn(bufferListener);
 
             player = new AndroidMediaPlayerImpl(mediaPlayer, listenersHolder, loadTimeout, heart, handler, buggyVideoDriverPreventer);
         }
